ADH-7835: Implement logic to skip database drop if it exists in source; add unit tests

+++ smart-hive-support/src/main/java/org/smartdata/hive/action/db/HmsDropDbAction.java
@@ -17,11 +17,20 @@
  */
 package org.smartdata.hive.action.db;
 
+import org.apache.commons.lang3.StringUtils;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hive.metastore.IMetaStoreClient;
+import org.apache.hadoop.hive.metastore.api.Database;
+import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
+import org.apache.hadoop.hive.metastore.conf.MetastoreConf;
 import org.apache.hadoop.hive.metastore.messaging.DropDatabaseMessage;
 import org.apache.hadoop.hive.metastore.messaging.EventMessage;
 import org.smartdata.action.annotation.ActionSignature;
+import org.smartdata.hdfs.impersonation.DisabledUserImpersonationStrategy;
+import org.smartdata.hive.HiveSmartConf;
 import org.smartdata.hive.action.HmsAction;
 import org.smartdata.hive.action.constraint.HmsCreateConstraintAction;
+import org.smartdata.hive.client.CachingMetaStoreClientProvider;
 
 @ActionSignature(
     actionId = HmsDropDbAction.NAME,
@@ -39,6 +48,12 @@ public class HmsDropDbAction extends HmsAction {
         EventMessage.EventType.DROP_DATABASE);
     appendFormatLog("Dropping database %s", message.getDB());
 
+    if (sourceDatabaseExists(message)) {
+      appendFormatLog("Skipping database drop on destination because '%s' still exists on source",
+          message.getDB());
+      return;
+    }
+
     getMetastoreClient().dropDatabase(
         message.getDB(),
         // deleteData
@@ -50,5 +65,45 @@ public class HmsDropDbAction extends HmsAction {
     );
 
     appendLog("Database was successfully dropped");
+  }
+
+  protected boolean sourceDatabaseExists(DropDatabaseMessage message) throws Exception {
+    try (IMetaStoreClient sourceMetastoreClient = sourceMetastoreClient()) {
+      return sourceDatabaseExists(sourceMetastoreClient, message);
+    }
+  }
+
+  private boolean sourceDatabaseExists(
+      IMetaStoreClient sourceMetastoreClient, DropDatabaseMessage message) throws Exception {
+    Database database = message.getDatabaseObject();
+    String databaseName = database == null ? message.getDB() : database.getName();
+    if (StringUtils.isBlank(databaseName)) {
+      throw new IllegalArgumentException("Source event contains empty database name");
+    }
+
+    try {
+      if (database == null || StringUtils.isBlank(database.getCatalogName())) {
+        sourceMetastoreClient.getDatabase(databaseName);
+      } else {
+        sourceMetastoreClient.getDatabase(database.getCatalogName(), databaseName);
+      }
+      return true;
+    } catch (NoSuchObjectException e) {
+      return false;
+    }
+  }
+
+  private IMetaStoreClient sourceMetastoreClient() {
+    HiveSmartConf hiveSmartConf = new HiveSmartConf(getContext().getConf());
+    Configuration metastoreConf = MetastoreConf.newMetastoreConf(hiveSmartConf);
+    String sourceMetastoreAddress = MetastoreConf.getVar(
+        metastoreConf, MetastoreConf.ConfVars.THRIFT_URIS);
+    if (StringUtils.isBlank(sourceMetastoreAddress)) {
+      throw new IllegalArgumentException("No source metastore address is configured");
+    }
+
+    return new CachingMetaStoreClientProvider(
+        hiveSmartConf, new DisabledUserImpersonationStrategy())
+        .provide(sourceMetastoreAddress, null);
   }
 }
